move tests added for this widget to be distributed with the project
