+++ phonenumber_field/tests/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from phonenumber_field.modelfields import PhoneNumberField
+from phonenumber_field.fields.models.caseinsensitivecharfield import CaseInsensitiveCharField
 
 ###############
 # Test Models #
@@ -11,3 +12,7 @@ class MandatoryPhoneNumber(models.Model):
 
 class OptionalPhoneNumber(models.Model):
     phone_number = PhoneNumberField(blank=True, default='')
+
+
+class CICharFieldTestModel(models.Model):
+    value = CaseInsensitiveCharField(primary_key=True, max_length=1)
