+++ testproject/testapp/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 from phonenumber_field.modelfields import PhoneNumberField
-from phonenumber_field.fields.models.caseinsensitivecharfield import CaseInsensitiveCharField
+
 
 # Create your models here.
 
@@ -13,5 +13,3 @@ class TestModelBlankPhone(models.Model):
     name = models.CharField(max_length=255, blank=True, default='')
     phone = PhoneNumberField(blank=True)
 
-class CICharFieldTestModel(models.Model):
-    value = CaseInsensitiveCharField(primary_key=True, max_length=1)
