+++ testproject/testapp/tests.py
@@ -6,12 +6,7 @@ Replace this with more appropriate tests for your application.
 """
 
 from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.db import IntegrityError, DEFAULT_DB_ALIAS
 from django.test import TestCase
-from unittest import skipIf
-
-DATABASE_IS_SQLITE = settings.DATABASES[DEFAULT_DB_ALIAS]['ENGINE'] == 'django.db.backends.sqlite3'
 
 class PhonenumerFieldAppTest(TestCase):
     def test_to_python_country_id_parse(self):
@@ -59,59 +54,3 @@ class PhonenumerFieldAppTest(TestCase):
         pk = tm.id
         tm = TestModelBlankPhone.objects.get(pk=pk)
         self.assertIsNone(tm.phone)
-
-class CICharFieldTestModelTestCase(TestCase):
-    def test_integrity_error(self):
-        from testapp.models import CICharFieldTestModel
-        self.assertEqual(CICharFieldTestModel.objects.count(), 0)
-        
-        CICharFieldTestModel.objects.create(value="a")
-        self.assertEqual(CICharFieldTestModel.objects.count(), 1)
-        
-        with self.assertRaises(IntegrityError):
-            CICharFieldTestModel.objects.create(value="A")
-    
-    def test_max_length(self):
-        from testapp.models import CICharFieldTestModel
-        obj = CICharFieldTestModel(value="bb")
-        with self.assertRaises(ValidationError):
-            obj.full_clean()
-    
-    @skipIf(DATABASE_IS_SQLITE, "Sqlite does not enforce varchar max_length")
-    def test_max_length_db(self):
-        from testapp.models import CICharFieldTestModel
-        self.assertEqual(CICharFieldTestModel.objects.count(), 0)
-        
-        CICharFieldTestModel.objects.create(value="bb")
-        self.assertNotEqual(CICharFieldTestModel.objects.all()[0].value.lower(), "bb")
-    
-    @skipIf(DATABASE_IS_SQLITE, "Sqlite does not enforce varchar max_length")
-    def test_max_length_db_truncates(self):
-        from testapp.models import CICharFieldTestModel
-        self.assertEqual(CICharFieldTestModel.objects.count(), 0)
-        
-        CICharFieldTestModel.objects.create(value="bb")
-        self.assertEqual(CICharFieldTestModel.objects.all()[0].value.lower(), "b")
-    
-    def test_lookup(self):
-        from testapp.models import CICharFieldTestModel
-        self.assertEqual(CICharFieldTestModel.objects.count(), 0)
-        
-        a_lower = "a"
-        a_upper = "A"
-        
-        CICharFieldTestModel.objects.create(value=a_lower)
-        self.assertEqual(CICharFieldTestModel.objects.count(), 1)
-        
-        a = CICharFieldTestModel.objects.get(value=a_upper)
-        self.assertEqual(a.value.upper(), a_upper)
-        
-        a.delete()
-        
-        self.assertEqual(CICharFieldTestModel.objects.count(), 0)
-        
-        CICharFieldTestModel.objects.create(value=a_upper)
-        self.assertEqual(CICharFieldTestModel.objects.count(), 1)
-        
-        A = CICharFieldTestModel.objects.get(value=a_lower)
-        self.assertEqual(A.value.lower(), a_lower)
